backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
websockets==12.0
groq==0.9.0
openai-whisper==20231117
soundfile==0.12.1
numpy>=1.21.0,<2.0.0
librosa==0.10.1
requests==2.31.0
python-dotenv==1.0.1
pydantic==2.7.1
setuptools
